+++ finance/src/test/kotlin/net/corda/contracts/asset/CashTests.kt
@@ -11,7 +11,7 @@ import net.corda.core.transactions.WireTransaction
 import net.corda.core.utilities.*
 import net.corda.node.services.vault.NodeVaultService
 import net.corda.node.utilities.configureDatabase
-import net.corda.node.utilities.databaseTransaction
+import net.corda.node.utilities.transaction
 import net.corda.testing.*
 import net.corda.testing.node.MockKeyManagementService
 import net.corda.testing.node.MockServices
@@ -53,7 +53,7 @@ class CashTests {
         val dataSourceAndDatabase = configureDatabase(dataSourceProps)
         dataSource = dataSourceAndDatabase.first
         database = dataSourceAndDatabase.second
-        databaseTransaction(database) {
+        database.transaction {
             services = object : MockServices() {
                 override val keyManagementService: MockKeyManagementService = MockKeyManagementService(MINI_CORP_KEY, MEGA_CORP_KEY, OUR_KEY)
                 override val vaultService: VaultService = makeVaultService(dataSourceProps)
@@ -484,7 +484,7 @@ class CashTests {
 
     fun makeSpend(amount: Amount<Currency>, dest: PublicKey): WireTransaction {
         val tx = TransactionType.General.Builder(DUMMY_NOTARY)
-        databaseTransaction(database) {
+        database.transaction {
             vault.generateSpend(tx, amount, dest)
         }
         return tx.toWireTransaction()
@@ -562,7 +562,7 @@ class CashTests {
     @Test
     fun generateSimpleDirectSpend() {
 
-        databaseTransaction(database) {
+        database.transaction {
 
             val wtx = makeSpend(100.DOLLARS, THEIR_PUBKEY_1)
 
@@ -577,7 +577,7 @@ class CashTests {
     @Test
     fun generateSimpleSpendWithParties() {
 
-        databaseTransaction(database) {
+        database.transaction {
 
             val tx = TransactionType.General.Builder(DUMMY_NOTARY)
             vault.generateSpend(tx, 80.DOLLARS, ALICE_PUBKEY, setOf(MINI_CORP.toAnonymous()))
@@ -589,7 +589,7 @@ class CashTests {
     @Test
     fun generateSimpleSpendWithChange() {
 
-        databaseTransaction(database) {
+        database.transaction {
 
             val wtx = makeSpend(10.DOLLARS, THEIR_PUBKEY_1)
 
@@ -605,7 +605,7 @@ class CashTests {
     @Test
     fun generateSpendWithTwoInputs() {
 
-        databaseTransaction(database) {
+        database.transaction {
             val wtx = makeSpend(500.DOLLARS, THEIR_PUBKEY_1)
 
             @Suppress("UNCHECKED_CAST")
@@ -621,7 +621,7 @@ class CashTests {
     @Test
     fun generateSpendMixedDeposits() {
 
-        databaseTransaction(database) {
+        database.transaction {
             val wtx = makeSpend(580.DOLLARS, THEIR_PUBKEY_1)
             assertEquals(3, wtx.inputs.size)
 
@@ -642,7 +642,7 @@ class CashTests {
     @Test
     fun generateSpendInsufficientBalance() {
 
-        databaseTransaction(database) {
+        database.transaction {
 
             val e: InsufficientBalanceException = assertFailsWith("balance") {
                 makeSpend(1000.DOLLARS, THEIR_PUBKEY_1)