+++ node/src/integration-test/kotlin/net/corda/node/services/RaftNotaryServiceTests.kt
@@ -13,7 +13,7 @@ import net.corda.flows.NotaryError
 import net.corda.flows.NotaryException
 import net.corda.flows.NotaryFlow
 import net.corda.node.internal.AbstractNode
-import net.corda.node.utilities.databaseTransaction
+import net.corda.node.utilities.transaction
 import net.corda.testing.node.NodeBasedTest
 import org.junit.Test
 import java.security.KeyPair
@@ -32,8 +32,8 @@ class RaftNotaryServiceTests : NodeBasedTest() {
         ).getOrThrow()
 
         val notaryParty = alice.netMapCache.getNotary(notaryName)!!
-        val notaryNodeKeyPair = databaseTransaction(masterNode.database) { masterNode.services.notaryIdentityKey }
-        val aliceKey = databaseTransaction(alice.database) { alice.services.legalIdentityKey }
+        val notaryNodeKeyPair = with(masterNode) { database.transaction { services.notaryIdentityKey } }
+        val aliceKey = with(alice) { database.transaction { services.legalIdentityKey } }
 
         val inputState = issueState(alice, notaryParty, notaryNodeKeyPair)
 
@@ -58,7 +58,7 @@ class RaftNotaryServiceTests : NodeBasedTest() {
     }
 
     private fun issueState(node: AbstractNode, notary: Party, notaryKey: KeyPair): StateAndRef<*> {
-        return databaseTransaction(node.database) {
+        return node.database.transaction {
             val tx = DummyContract.generateInitial(Random().nextInt(), notary, node.info.legalIdentity.ref(0))
             tx.signWith(node.services.legalIdentityKey)
             tx.signWith(notaryKey)