+++ node/src/main/kotlin/net/corda/node/internal/CordaRPCOpsImpl.kt
@@ -22,7 +22,7 @@ import net.corda.node.services.startFlowPermission
 import net.corda.node.services.statemachine.FlowStateMachineImpl
 import net.corda.node.services.statemachine.StateMachineManager
 import net.corda.node.utilities.AddOrRemove
-import net.corda.node.utilities.databaseTransaction
+import net.corda.node.utilities.transaction
 import org.jetbrains.exposed.sql.Database
 import rx.Observable
 import java.io.InputStream
@@ -42,26 +42,26 @@ class CordaRPCOpsImpl(
     override val protocolVersion: Int get() = 0
 
     override fun networkMapUpdates(): Pair<List<NodeInfo>, Observable<NetworkMapCache.MapChange>> {
-        return databaseTransaction(database) {
+        return database.transaction {
             services.networkMapCache.track()
         }
     }
 
     override fun vaultAndUpdates(): Pair<List<StateAndRef<ContractState>>, Observable<Vault.Update>> {
-        return databaseTransaction(database) {
+        return database.transaction {
             val (vault, updates) = services.vaultService.track()
             Pair(vault.states.toList(), updates)
         }
     }
 
     override fun verifiedTransactions(): Pair<List<SignedTransaction>, Observable<SignedTransaction>> {
-        return databaseTransaction(database) {
+        return database.transaction {
             services.storageService.validatedTransactions.track()
         }
     }
 
     override fun stateMachinesAndUpdates(): Pair<List<StateMachineInfo>, Observable<StateMachineUpdate>> {
-        return databaseTransaction(database) {
+        return database.transaction {
             val (allStateMachines, changes) = smm.track()
             Pair(
                     allStateMachines.map { stateMachineInfoFromFlowLogic(it.id, it.logic) },
@@ -71,7 +71,7 @@ class CordaRPCOpsImpl(
     }
 
     override fun stateMachineRecordedTransactionMapping(): Pair<List<StateMachineTransactionMapping>, Observable<StateMachineTransactionMapping>> {
-        return databaseTransaction(database) {
+        return database.transaction {
             services.storageService.stateMachineRecordedTransactionMapping.track()
         }
     }
@@ -81,19 +81,19 @@ class CordaRPCOpsImpl(
     }
 
     override fun addVaultTransactionNote(txnId: SecureHash, txnNote: String) {
-        return databaseTransaction(database) {
+        return database.transaction {
             services.vaultService.addNoteToTransaction(txnId, txnNote)
         }
     }
 
     override fun getVaultTransactionNotes(txnId: SecureHash): Iterable<String> {
-        return databaseTransaction(database) {
+        return database.transaction {
             services.vaultService.getTransactionNotes(txnId)
         }
     }
 
     override fun getCashBalances(): Map<Currency, Amount<Currency>> {
-        return databaseTransaction(database) {
+        return database.transaction {
             services.vaultService.cashBalances
         }
     }
@@ -111,21 +111,21 @@ class CordaRPCOpsImpl(
 
     override fun attachmentExists(id: SecureHash): Boolean {
         // TODO: this operation should not require an explicit transaction
-        return databaseTransaction(database) {
+        return database.transaction {
             services.storageService.attachments.openAttachment(id) != null
         }
     }
 
     override fun openAttachment(id: SecureHash): InputStream {
         // TODO: this operation should not require an explicit transaction
-        return databaseTransaction(database) {
+        return database.transaction {
             services.storageService.attachments.openAttachment(id)!!.open()
         }
     }
 
     override fun uploadAttachment(jar: InputStream): SecureHash {
         // TODO: this operation should not require an explicit transaction
-        return databaseTransaction(database) {
+        return database.transaction {
             services.storageService.attachments.importAttachment(jar)
         }
     }
@@ -136,7 +136,7 @@ class CordaRPCOpsImpl(
     @Suppress("OverridingDeprecatedMember", "DEPRECATION")
     override fun uploadFile(dataType: String, name: String?, file: InputStream): String {
         val acceptor = services.storageService.uploaders.firstOrNull { it.accepts(dataType) }
-        return databaseTransaction(database) {
+        return database.transaction {
             acceptor?.upload(file) ?: throw RuntimeException("Cannot find file upload acceptor for $dataType")
         }
     }