+++ node/src/test/kotlin/net/corda/node/services/HibernateObserverTests.kt
@@ -14,7 +14,7 @@ import net.corda.core.utilities.LogHelper
 import net.corda.node.services.api.SchemaService
 import net.corda.node.services.schema.HibernateObserver
 import net.corda.node.utilities.configureDatabase
-import net.corda.node.utilities.databaseTransaction
+import net.corda.node.utilities.transaction
 import net.corda.testing.MEGA_CORP
 import net.corda.testing.node.makeTestDataSourceProperties
 import org.hibernate.annotations.Cascade
@@ -110,7 +110,7 @@ class HibernateObserverTests {
 
         @Suppress("UNUSED_VARIABLE")
         val observer = HibernateObserver(rawUpdatesPublisher, schemaService)
-        databaseTransaction(database) {
+        database.transaction {
             rawUpdatesPublisher.onNext(Vault.Update(emptySet(), setOf(StateAndRef(TransactionState(TestState(), MEGA_CORP), StateRef(SecureHash.sha256("dummy"), 0)))))
             val parentRowCountResult = TransactionManager.current().connection.prepareStatement("select count(*) from contract_Parents").executeQuery()
             parentRowCountResult.next()