+++ node/src/test/kotlin/net/corda/node/services/database/RequeryConfigurationTest.kt
@@ -23,7 +23,7 @@ import net.corda.node.services.vault.schemas.VaultCashBalancesEntity
 import net.corda.node.services.vault.schemas.VaultSchema
 import net.corda.node.services.vault.schemas.VaultStatesEntity
 import net.corda.node.utilities.configureDatabase
-import net.corda.node.utilities.databaseTransaction
+import net.corda.node.utilities.transaction
 import net.corda.testing.node.makeTestDataSourceProperties
 import org.assertj.core.api.Assertions
 import org.jetbrains.exposed.sql.Database
@@ -61,14 +61,14 @@ class RequeryConfigurationTest {
     fun `transaction inserts in same DB transaction scope across two persistence engines`() {
         val txn = newTransaction()
 
-        databaseTransaction(database) {
+        database.transaction {
             transactionStorage.addTransaction(txn)
             requerySession.withTransaction {
                 insert(createVaultStateEntity(txn))
             }
         }
 
-        databaseTransaction(database) {
+        database.transaction {
             Assertions.assertThat(transactionStorage.transactions).containsOnly(txn)
             requerySession.withTransaction {
                 val result = select(VaultSchema.VaultStates::class) where (VaultSchema.VaultStates::txId eq txn.tx.inputs[0].txhash.toString())
@@ -81,7 +81,7 @@ class RequeryConfigurationTest {
     fun `transaction operations in same DB transaction scope across two persistence engines`() {
         val txn = newTransaction()
 
-        databaseTransaction(database) {
+        database.transaction {
             transactionStorage.addTransaction(txn)
             requerySession.withTransaction {
                 upsert(createCashBalance())
@@ -90,7 +90,7 @@ class RequeryConfigurationTest {
             }
         }
 
-        databaseTransaction(database) {
+        database.transaction {
             Assertions.assertThat(transactionStorage.transactions).containsOnly(txn)
             requerySession.withTransaction {
                 val cashQuery = select(VaultSchema.VaultCashBalances::class) where (VaultSchema.VaultCashBalances::currency eq "GBP")
@@ -105,7 +105,7 @@ class RequeryConfigurationTest {
     fun `transaction rollback in same DB transaction scope across two persistence engines`() {
         val txn = newTransaction()
 
-        databaseTransaction(database) {
+        database.transaction {
             transactionStorage.addTransaction(txn)
             requerySession.withTransaction {
                 insert(createVaultStateEntity(txn))
@@ -113,7 +113,7 @@ class RequeryConfigurationTest {
             rollback()
         }
 
-        databaseTransaction(database) {
+        database.transaction {
             Assertions.assertThat(transactionStorage.transactions).isEmpty()
             requerySession.withTransaction {
                 val result = select(VaultSchema.VaultStates::class) where (VaultSchema.VaultStates::txId eq txn.tx.inputs[0].txhash.toString())
@@ -145,13 +145,13 @@ class RequeryConfigurationTest {
     }
 
     private fun newTransactionStorage() {
-        databaseTransaction(database) {
+        database.transaction {
             transactionStorage = DBTransactionStorage()
         }
     }
 
     private fun newRequeryStorage(dataSourceProperties: Properties) {
-        databaseTransaction(database) {
+        database.transaction {
             val configuration = RequeryConfiguration(dataSourceProperties, true)
             requerySession = configuration.sessionForModel(Models.VAULT)
         }